requirements.txt
torch
torchvision
Pillow
pandas
tqdm
scikit-learn
typing
argparse
matplotlib